+++ src/inspect_ai/model/_model_output.py
@@ -3,6 +3,7 @@ from typing import Literal
 from pydantic import BaseModel, Field
 
 from ._chat_message import ChatMessageAssistant
+from ._tool import ToolCall
 
 
 class ModelUsage(BaseModel):
@@ -121,4 +122,40 @@ class ModelOutput(BaseModel):
                 )
             ],
             error=error,
+        )
+
+    @staticmethod
+    def for_tool_call(
+        model: str, tool_name: str, tool_arguments: dict[str, str]
+    ) -> "ModelOutput":
+        """
+        Returns a ModelOutput for requesting a tool call.
+
+        Args:
+            model: model name
+            tool_name: The name of the tool.
+            tool_arguments: The arguments passed to the tool.
+
+        Returns:
+            A ModelOutput corresponding to the tool call
+        """
+        return ModelOutput(
+            model=model,
+            choices=[
+                ChatCompletionChoice(
+                    message=ChatMessageAssistant(
+                        content=f"tool call for tool {tool_name}",
+                        source="generate",
+                        tool_calls=[
+                            ToolCall(
+                                id="tool_call_id",
+                                function=tool_name,
+                                arguments=tool_arguments,
+                                type="function",
+                            )
+                        ],
+                    ),
+                    stop_reason="tool_calls",
+                )
+            ],
         )