+++ tools/vscode/src/providers/logview/logview-manager.ts
@@ -1,11 +1,17 @@
 import { Uri, ViewColumn, window, workspace } from "vscode";
 import { InspectLogviewWebviewManager } from "./logview-webview";
 import { InspectSettingsManager } from "../settings/inspect-settings";
+import { WorkspaceEnvManager } from "../workspace/workspace-env-provider";
+import { activeWorkspaceFolder } from "../../core/workspace";
+import { workspacePath } from "../../core/path";
+import { kInspectEnvValues } from "../inspect/inspect-constants";
+
 
 export class InspectLogviewManager {
   constructor(
     private readonly webViewManager_: InspectLogviewWebviewManager,
     private readonly settingsMgr_: InspectSettingsManager,
+    private readonly envMgr_: WorkspaceEnvManager
   ) { }
 
   public async showLogFile(logFile: Uri) {
@@ -24,7 +30,23 @@ export class InspectLogviewManager {
   }
 
   public showInspectView() {
-    this.webViewManager_.showLogview();
+
+    // See if there is a log dir
+    const envVals = this.envMgr_.getValues();
+    const env_log = envVals[kInspectEnvValues.logDir];
+
+    // If there is a log dir, try to parse and use it
+    let log_uri;
+    try {
+      log_uri = Uri.parse(env_log, true);
+    } catch {
+      // This isn't a uri, bud
+      log_uri = Uri.file(workspacePath(env_log).path);
+    }
+
+    // Show the log view for the log dir (or the workspace)
+    const log_dir = log_uri || activeWorkspaceFolder().uri;
+    this.webViewManager_.showLogview({ log_dir });
   }
 
   public viewColumn() {